Fixed incorrect type in loop

+++ examples/deserialize.c
@@ -16,7 +16,7 @@ int main(void) {
         cst_print_node(stdout, root->nodes[i]);
     }
     // Cleanup
-    for (int i = 0; i < root->n_nodes; i++) {
+    for (size_t i = 0; i < root->n_nodes; i++) {
         cst_free_node(root->nodes[i]);
         free(root->nodes[i]);
     }
